requirements.txt
kafka-python
requests
random
json
flask
flask_socketio
python-dotenv
os